Removing some compute ap bug

+++ free_loc/custom.py
@@ -244,8 +244,8 @@ def compute_ap(gt, pred, valid, average=None):
     nclasses = gt.shape[1]
     AP = []
     for cid in range(nclasses):
-        gt_cls = gt[:, cid][valid[:, cid] > 0].astype('float32')
-        pred_cls = pred[:, cid][valid[:, cid] > 0].astype('float32')
+        gt_cls = gt[:, cid][(valid[:, cid] > 0).astype(int)].astype('float32')
+        pred_cls = pred[:, cid][(valid[:, cid] > 0).astype(int)].astype('float32')
         # As per PhilK. code:
         # https://github.com/philkr/voc-classification/blob/master/src/train_cls.py
         pred_cls -= 1e-5 * gt_cls
